refactor: created common package for argo cli to make funcs public for archive
