+++ cmd/argo/commands/common/common.go
@@ -1,4 +1,4 @@
-package commands
+package common
 
 import (
 	"fmt"
@@ -14,11 +14,11 @@ import (
 
 // Global variables
 var (
-	jobStatusIconMap         map[wfv1.NodePhase]string
-	nodeTypeIconMap          map[wfv1.NodeType]string
-	workflowConditionIconMap map[wfv1.ConditionType]string
-	noColor                  bool
-	noUtf8                   bool
+	JobStatusIconMap         map[wfv1.NodePhase]string
+	NodeTypeIconMap          map[wfv1.NodeType]string
+	WorkflowConditionIconMap map[wfv1.ConditionType]string
+	NoColor                  bool
+	NoUtf8                   bool
 )
 
 func init() {
@@ -46,8 +46,8 @@ func initializeSession() {
 		TimestampFormat: "2006-01-02T15:04:05.000Z",
 		FullTimestamp:   true,
 	})
-	if noUtf8 {
-		jobStatusIconMap = map[wfv1.NodePhase]string{
+	if NoUtf8 {
+		JobStatusIconMap = map[wfv1.NodePhase]string{
 			wfv1.NodePending:   ansiFormat("Pending", FgYellow),
 			wfv1.NodeRunning:   ansiFormat("Running", FgCyan),
 			wfv1.NodeSucceeded: ansiFormat("Succeeded", FgGreen),
@@ -55,15 +55,15 @@ func initializeSession() {
 			wfv1.NodeFailed:    ansiFormat("Failed", FgRed),
 			wfv1.NodeError:     ansiFormat("Error", FgRed),
 		}
-		nodeTypeIconMap = map[wfv1.NodeType]string{
+		NodeTypeIconMap = map[wfv1.NodeType]string{
 			wfv1.NodeTypeSuspend: ansiFormat("Suspend", FgCyan),
 		}
-		workflowConditionIconMap = map[wfv1.ConditionType]string{
+		WorkflowConditionIconMap = map[wfv1.ConditionType]string{
 			wfv1.ConditionTypeMetricsError: ansiFormat("Error", FgRed),
 			wfv1.ConditionTypeSpecWarning:  ansiFormat("Warning", FgYellow),
 		}
 	} else {
-		jobStatusIconMap = map[wfv1.NodePhase]string{
+		JobStatusIconMap = map[wfv1.NodePhase]string{
 			wfv1.NodePending:   ansiFormat("◷", FgYellow),
 			wfv1.NodeRunning:   ansiFormat("●", FgCyan),
 			wfv1.NodeSucceeded: ansiFormat("✔", FgGreen),
@@ -71,10 +71,10 @@ func initializeSession() {
 			wfv1.NodeFailed:    ansiFormat("✖", FgRed),
 			wfv1.NodeError:     ansiFormat("⚠", FgRed),
 		}
-		nodeTypeIconMap = map[wfv1.NodeType]string{
+		NodeTypeIconMap = map[wfv1.NodeType]string{
 			wfv1.NodeTypeSuspend: ansiFormat("ǁ", FgCyan),
 		}
-		workflowConditionIconMap = map[wfv1.ConditionType]string{
+		WorkflowConditionIconMap = map[wfv1.ConditionType]string{
 			wfv1.ConditionTypeMetricsError: ansiFormat("✖", FgRed),
 			wfv1.ConditionTypeSpecWarning:  ansiFormat("⚠", FgYellow),
 		}
@@ -98,7 +98,7 @@ func ansiColorCode(s string) int {
 // color, it provides more consistent string lengths so that tabwriter can calculate
 // widths correctly.
 func ansiFormat(s string, codes ...int) string {
-	if noColor || os.Getenv("TERM") == "dumb" || len(codes) == 0 {
+	if NoColor || os.Getenv("TERM") == "dumb" || len(codes) == 0 {
 		return s
 	}
 	codeStrs := make([]string, len(codes))
