+++ cmd/argo/commands/common/common_test.go
@@ -1,4 +1,4 @@
-package commands
+package common
 
 import (
 	"testing"
