+++ cmd/argo/commands/resubmit_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/argoproj/argo-workflows/v3/cmd/argo/commands/common"
 	workflowpkg "github.com/argoproj/argo-workflows/v3/pkg/apiclient/workflow"
 	workflowmocks "github.com/argoproj/argo-workflows/v3/pkg/apiclient/workflow/mocks"
 	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
@@ -20,7 +21,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 		resubmitOpts := resubmitOps{
 			namespace: "argo",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		c.On("ResubmitWorkflow", mock.Anything, mock.Anything).Return(&wfv1.Workflow{}, nil)
 
@@ -36,7 +37,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 			namespace: "argo",
 			memoized:  true,
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		c.On("ResubmitWorkflow", mock.Anything, mock.Anything).Return(&wfv1.Workflow{}, nil)
 
@@ -57,7 +58,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		wfListReq := &workflowpkg.WorkflowListRequest{
 			Namespace: "argo",
@@ -97,7 +98,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		wfListReq := &workflowpkg.WorkflowListRequest{
 			Namespace: "argo",
@@ -147,7 +148,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 		c.On("ListWorkflows", mock.Anything, mock.Anything).Return(nil, fmt.Errorf("mock error"))
 		err := resubmitWorkflows(context.Background(), c, resubmitOpts, cliSubmitOpts, []string{})
 		assert.Errorf(t, err, "mock error")
@@ -158,7 +159,7 @@ func Test_resubmitWorkflows(t *testing.T) {
 		resubmitOpts := resubmitOps{
 			namespace: "argo",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 		c.On("ResubmitWorkflow", mock.Anything, mock.Anything).Return(nil, fmt.Errorf("mock error"))
 		err := resubmitWorkflows(context.Background(), c, resubmitOpts, cliSubmitOpts, []string{"foo"})
 		assert.Errorf(t, err, "mock error")
