+++ cmd/argo/commands/retry_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/argoproj/argo-workflows/v3/cmd/argo/commands/common"
 	workflowpkg "github.com/argoproj/argo-workflows/v3/pkg/apiclient/workflow"
 	workflowmocks "github.com/argoproj/argo-workflows/v3/pkg/apiclient/workflow/mocks"
 	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
@@ -20,7 +21,7 @@ func Test_retryWorkflows(t *testing.T) {
 		retryOpts := retryOps{
 			namespace: "argo",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		c.On("RetryWorkflow", mock.Anything, mock.Anything).Return(&wfv1.Workflow{}, nil)
 
@@ -36,7 +37,7 @@ func Test_retryWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		wfListReq := &workflowpkg.WorkflowListRequest{
 			Namespace: "argo",
@@ -77,7 +78,7 @@ func Test_retryWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 
 		wfListReq := &workflowpkg.WorkflowListRequest{
 			Namespace: "argo",
@@ -129,7 +130,7 @@ func Test_retryWorkflows(t *testing.T) {
 			namespace:     "argo",
 			labelSelector: "custom-label=true",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 		c.On("ListWorkflows", mock.Anything, mock.Anything).Return(nil, fmt.Errorf("mock error"))
 		err := retryWorkflows(context.Background(), c, retryOpts, cliSubmitOpts, []string{})
 		assert.Errorf(t, err, "mock error")
@@ -140,7 +141,7 @@ func Test_retryWorkflows(t *testing.T) {
 		retryOpts := retryOps{
 			namespace: "argo",
 		}
-		cliSubmitOpts := cliSubmitOpts{}
+		cliSubmitOpts := common.CliSubmitOpts{}
 		c.On("RetryWorkflow", mock.Anything, mock.Anything).Return(nil, fmt.Errorf("mock error"))
 		err := retryWorkflows(context.Background(), c, retryOpts, cliSubmitOpts, []string{"foo"})
 		assert.Errorf(t, err, "mock error")
