+++ cmd/argo/commands/submit.go
@@ -11,32 +11,20 @@ import (
 	argoJson "github.com/argoproj/pkg/json"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/argoproj/argo-workflows/v3/cmd/argo/commands/client"
+	common "github.com/argoproj/argo-workflows/v3/cmd/argo/commands/common"
 	workflowpkg "github.com/argoproj/argo-workflows/v3/pkg/apiclient/workflow"
 	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
-	"github.com/argoproj/argo-workflows/v3/workflow/common"
+	wfcommon "github.com/argoproj/argo-workflows/v3/workflow/common"
 	"github.com/argoproj/argo-workflows/v3/workflow/util"
 )
-
-// cliSubmitOpts holds submission options specific to CLI submission (e.g. controlling output)
-type cliSubmitOpts struct {
-	output        string // --output
-	wait          bool   // --wait
-	watch         bool   // --watch
-	log           bool   // --log
-	strict        bool   // --strict
-	priority      *int32 // --priority
-	getArgs       getFlags
-	scheduledTime string // --scheduled-time
-}
 
 func NewSubmitCommand() *cobra.Command {
 	var (
 		submitOpts    wfv1.SubmitOpts
-		cliSubmitOpts cliSubmitOpts
+		cliSubmitOpts common.CliSubmitOpts
 		priority      int32
 		from          string
 	)
@@ -65,10 +53,10 @@ func NewSubmitCommand() *cobra.Command {
 `,
 		Run: func(cmd *cobra.Command, args []string) {
 			if cmd.Flag("priority").Changed {
-				cliSubmitOpts.priority = &priority
+				cliSubmitOpts.Priority = &priority
 			}
 
-			if !cliSubmitOpts.watch && len(cliSubmitOpts.getArgs.status) > 0 {
+			if !cliSubmitOpts.Watch && len(cliSubmitOpts.GetArgs.Status) > 0 {
 				log.Warn("--status should only be used with --watch")
 			}
 
@@ -87,16 +75,16 @@ func NewSubmitCommand() *cobra.Command {
 		},
 	}
 	util.PopulateSubmitOpts(command, &submitOpts, true)
-	command.Flags().StringVarP(&cliSubmitOpts.output, "output", "o", "", "Output format. One of: name|json|yaml|wide")
-	command.Flags().BoolVarP(&cliSubmitOpts.wait, "wait", "w", false, "wait for the workflow to complete")
-	command.Flags().BoolVar(&cliSubmitOpts.watch, "watch", false, "watch the workflow until it completes")
-	command.Flags().BoolVar(&cliSubmitOpts.log, "log", false, "log the workflow until it completes")
-	command.Flags().BoolVar(&cliSubmitOpts.strict, "strict", true, "perform strict workflow validation")
+	command.Flags().StringVarP(&cliSubmitOpts.Output, "output", "o", "", "Output format. One of: name|json|yaml|wide")
+	command.Flags().BoolVarP(&cliSubmitOpts.Wait, "wait", "w", false, "wait for the workflow to complete")
+	command.Flags().BoolVar(&cliSubmitOpts.Watch, "watch", false, "watch the workflow until it completes")
+	command.Flags().BoolVar(&cliSubmitOpts.Log, "log", false, "log the workflow until it completes")
+	command.Flags().BoolVar(&cliSubmitOpts.Strict, "strict", true, "perform strict workflow validation")
 	command.Flags().Int32Var(&priority, "priority", 0, "workflow priority")
 	command.Flags().StringVar(&from, "from", "", "Submit from an existing `kind/name` E.g., --from=cronwf/hello-world-cwf")
-	command.Flags().StringVar(&cliSubmitOpts.getArgs.status, "status", "", "Filter by status (Pending, Running, Succeeded, Skipped, Failed, Error). Should only be used with --watch.")
-	command.Flags().StringVar(&cliSubmitOpts.getArgs.nodeFieldSelectorString, "node-field-selector", "", "selector of node to display, eg: --node-field-selector phase=abc")
-	command.Flags().StringVar(&cliSubmitOpts.scheduledTime, "scheduled-time", "", "Override the workflow's scheduledTime parameter (useful for backfilling). The time must be RFC3339")
+	command.Flags().StringVar(&cliSubmitOpts.GetArgs.Status, "status", "", "Filter by status (Pending, Running, Succeeded, Skipped, Failed, Error). Should only be used with --watch.")
+	command.Flags().StringVar(&cliSubmitOpts.GetArgs.NodeFieldSelectorString, "node-field-selector", "", "selector of node to display, eg: --node-field-selector phase=abc")
+	command.Flags().StringVar(&cliSubmitOpts.ScheduledTime, "scheduled-time", "", "Override the workflow's scheduledTime parameter (useful for backfilling). The time must be RFC3339")
 
 	// Only complete files with appropriate extension.
 	err := command.Flags().SetAnnotation("parameter-file", cobra.BashCompFilenameExt, []string{"json", "yaml", "yml"})
@@ -106,25 +94,25 @@ func NewSubmitCommand() *cobra.Command {
 	return command
 }
 
-func submitWorkflowsFromFile(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, filePaths []string, submitOpts *wfv1.SubmitOpts, cliOpts *cliSubmitOpts) {
+func submitWorkflowsFromFile(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, filePaths []string, submitOpts *wfv1.SubmitOpts, cliOpts *common.CliSubmitOpts) {
 	fileContents, err := util.ReadManifest(filePaths...)
 	errors.CheckError(err)
 
 	var workflows []wfv1.Workflow
 	for _, body := range fileContents {
-		wfs := unmarshalWorkflows(body, cliOpts.strict)
+		wfs := unmarshalWorkflows(body, cliOpts.Strict)
 		workflows = append(workflows, wfs...)
 	}
 
 	submitWorkflows(ctx, serviceClient, namespace, workflows, submitOpts, cliOpts)
 }
 
-func validateOptions(workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cliOpts *cliSubmitOpts) {
-	if cliOpts.watch {
+func validateOptions(workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cliOpts *common.CliSubmitOpts) {
+	if cliOpts.Watch {
 		if len(workflows) > 1 {
 			log.Fatalf("Cannot watch more than one workflow")
 		}
-		if cliOpts.wait {
+		if cliOpts.Wait {
 			log.Fatalf("--wait cannot be combined with --watch")
 		}
 		if submitOpts.DryRun {
@@ -135,7 +123,7 @@ func validateOptions(workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cli
 		}
 	}
 
-	if cliOpts.wait {
+	if cliOpts.Wait {
 		if submitOpts.DryRun {
 			log.Fatalf("--wait cannot be combined with --dry-run")
 		}
@@ -145,7 +133,7 @@ func validateOptions(workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cli
 	}
 
 	if submitOpts.DryRun {
-		if cliOpts.output == "" {
+		if cliOpts.Output == "" {
 			log.Fatalf("--dry-run should have an output option")
 		}
 		if submitOpts.ServerDryRun {
@@ -154,13 +142,13 @@ func validateOptions(workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cli
 	}
 
 	if submitOpts.ServerDryRun {
-		if cliOpts.output == "" {
+		if cliOpts.Output == "" {
 			log.Fatalf("--server-dry-run should have an output option")
 		}
 	}
 }
 
-func submitWorkflowFromResource(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, resourceIdentifier string, submitOpts *wfv1.SubmitOpts, cliOpts *cliSubmitOpts) {
+func submitWorkflowFromResource(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, resourceIdentifier string, submitOpts *wfv1.SubmitOpts, cliOpts *common.CliSubmitOpts) {
 	parts := strings.SplitN(resourceIdentifier, "/", 2)
 	if len(parts) != 2 {
 		log.Fatalf("resource identifier '%s' is malformed. Should be `kind/name`, e.g. cronwf/hello-world-cwf", resourceIdentifier)
@@ -171,12 +159,12 @@ func submitWorkflowFromResource(ctx context.Context, serviceClient workflowpkg.W
 	tempwf := wfv1.Workflow{}
 
 	validateOptions([]wfv1.Workflow{tempwf}, submitOpts, cliOpts)
-	if cliOpts.scheduledTime != "" {
-		_, err := time.Parse(time.RFC3339, cliOpts.scheduledTime)
+	if cliOpts.ScheduledTime != "" {
+		_, err := time.Parse(time.RFC3339, cliOpts.ScheduledTime)
 		if err != nil {
 			log.Fatalf("scheduled-time contains invalid time.RFC3339 format. (e.g.: `2006-01-02T15:04:05-07:00`)")
 		}
-		submitOpts.Annotations = fmt.Sprintf("%s=%s", common.AnnotationKeyCronWfScheduledTime, cliOpts.scheduledTime)
+		submitOpts.Annotations = fmt.Sprintf("%s=%s", wfcommon.AnnotationKeyCronWfScheduledTime, cliOpts.ScheduledTime)
 	}
 
 	created, err := serviceClient.SubmitWorkflow(ctx, &workflowpkg.WorkflowSubmitRequest{
@@ -189,12 +177,12 @@ func submitWorkflowFromResource(ctx context.Context, serviceClient workflowpkg.W
 		log.Fatalf("Failed to submit workflow: %v", err)
 	}
 
-	printWorkflow(created, getFlags{output: cliOpts.output})
+	printWorkflow(created, common.GetFlags{Output: cliOpts.Output})
 
-	waitWatchOrLog(ctx, serviceClient, namespace, []string{created.Name}, *cliOpts)
+	common.WaitWatchOrLog(ctx, serviceClient, namespace, []string{created.Name}, *cliOpts)
 }
 
-func submitWorkflows(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cliOpts *cliSubmitOpts) {
+func submitWorkflows(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, workflows []wfv1.Workflow, submitOpts *wfv1.SubmitOpts, cliOpts *common.CliSubmitOpts) {
 	validateOptions(workflows, submitOpts, cliOpts)
 
 	if len(workflows) == 0 {
@@ -211,7 +199,7 @@ func submitWorkflows(ctx context.Context, serviceClient workflowpkg.WorkflowServ
 		}
 		err := util.ApplySubmitOpts(&wf, submitOpts)
 		errors.CheckError(err)
-		wf.Spec.Priority = cliOpts.priority
+		wf.Spec.Priority = cliOpts.Priority
 		options := &metav1.CreateOptions{}
 		if submitOpts.DryRun {
 			options.DryRun = []string{"All"}
@@ -226,11 +214,11 @@ func submitWorkflows(ctx context.Context, serviceClient workflowpkg.WorkflowServ
 			log.Fatalf("Failed to submit workflow: %v", err)
 		}
 
-		printWorkflow(created, getFlags{output: cliOpts.output, status: cliOpts.getArgs.status})
+		printWorkflow(created, common.GetFlags{Output: cliOpts.Output, Status: cliOpts.GetArgs.Status})
 		workflowNames = append(workflowNames, created.Name)
 	}
 
-	waitWatchOrLog(ctx, serviceClient, namespace, workflowNames, *cliOpts)
+	common.WaitWatchOrLog(ctx, serviceClient, namespace, workflowNames, *cliOpts)
 }
 
 // unmarshalWorkflows unmarshals the input bytes as either json or yaml
@@ -244,29 +232,10 @@ func unmarshalWorkflows(wfBytes []byte, strict bool) []wfv1.Workflow {
 	if err == nil {
 		return []wfv1.Workflow{wf}
 	}
-	yamlWfs, err := common.SplitWorkflowYAMLFile(wfBytes, strict)
+	yamlWfs, err := wfcommon.SplitWorkflowYAMLFile(wfBytes, strict)
 	if err == nil {
 		return yamlWfs
 	}
 	log.Fatalf("Failed to parse workflow: %v", err)
 	return nil
-}
-
-func waitWatchOrLog(ctx context.Context, serviceClient workflowpkg.WorkflowServiceClient, namespace string, workflowNames []string, cliSubmitOpts cliSubmitOpts) {
-	if cliSubmitOpts.log {
-		for _, workflow := range workflowNames {
-			logWorkflow(ctx, serviceClient, namespace, workflow, "", "", "", &corev1.PodLogOptions{
-				Container: common.MainContainerName,
-				Follow:    true,
-				Previous:  false,
-			})
-		}
-	}
-	if cliSubmitOpts.wait {
-		waitWorkflows(ctx, serviceClient, namespace, workflowNames, false, !(cliSubmitOpts.output == "" || cliSubmitOpts.output == "wide"))
-	} else if cliSubmitOpts.watch {
-		for _, workflow := range workflowNames {
-			watchWorkflow(ctx, serviceClient, namespace, workflow, cliSubmitOpts.getArgs)
-		}
-	}
 }
